+++ index.js
@@ -13,7 +13,6 @@ export {
 
 export {
 	wasInstantNavigation,
-	isInstantBackAbleNavigation,
 	canGoBackInstantly,
 	canGoForwardInstantly
 } from './lib/redux/client/instantNavigation.js'
@@ -50,10 +49,16 @@ export {
 	goForward
 } from './lib/router/actions.js'
 
-export { default as useNavigationStartEffect } from './lib/router/useNavigationStartEffect.js'
-export { default as useNavigationEndEffect } from './lib/router/useNavigationEndEffect.js'
+export { default as useBeforeNavigateToAnotherPage } from './lib/router/useBeforeNavigateToAnotherPage.js'
+export { default as useBeforeRenderAnotherPage } from './lib/router/useBeforeRenderAnotherPage.js'
+// export { default as useAfterNavigatedToAnotherPage } from './lib/router/useAfterNavigatedToAnotherPage.js'
+export { default as useAfterRenderedThisPage } from './lib/router/useAfterRenderedThisPage.js'
+export { default as useBeforeRenderNewPage } from './lib/router/useBeforeRenderNewPage.js'
+export { default as useAfterRenderedNewPage } from './lib/router/useAfterRenderedNewPage.js'
 export { default as useNavigationLocation } from './lib/redux/navigation/useNavigationLocation.js'
-export { default as useSelectorForLocation } from './lib/redux/useSelectorForLocation.js'
+// export { default as useSelectorForLocation } from './lib/redux/useSelectorForLocation.js'
+export { default as usePageStateSelector } from './lib/redux/usePageStateSelector.js'
+export { default as usePageStateSelectorOutsideOfPage } from './lib/redux/usePageStateSelectorOutsideOfPage.js'
 export { default as useLocation } from './lib/router/useLocation.js'
 export { default as useLocationHistory } from './lib/router/useLocationHistory.js'
 export { default as useGoBack } from './lib/router/useGoBack.js'