+++ types.d.ts
@@ -27,7 +27,7 @@ interface HttpClient {
 // `superagent` request.
 type HttpRequest_ = object;
 
-interface RouteWithOrWithoutCodeSplit<State, Action extends ReduxAction<string>, Context> {
+interface RouteWithOrWithoutCodeSplit<State, Action extends ReduxAction<string>, LoadContext> {
 	path?: string;
 
 	Component?: React.FC;
@@ -38,9 +38,9 @@ interface RouteWithOrWithoutCodeSplit<State, Action extends ReduxAction<string>,
 	load?: (parameters: {
 		dispatch: Dispatch<Action>,
 		useSelector: TypedUseSelectorHook<State>,
-		params: object,
+		params: Record<string, string>,
 		location: Location,
-		context?: Context,
+		context?: LoadContext,
 		history: {
 			route: string,
 			action: string
@@ -49,11 +49,11 @@ interface RouteWithOrWithoutCodeSplit<State, Action extends ReduxAction<string>,
 		getCookie: (name: string) => string | undefined
 	}) => Promise<void>;
 
-	children?: RouteWithOrWithoutCodeSplit<State, Action, Context>[];
+	children?: RouteWithOrWithoutCodeSplit<State, Action, LoadContext>[];
 }
 
-export interface Settings<State, Action extends ReduxAction<string>, Context> {
-	routes: RouteWithOrWithoutCodeSplit<State, Action, Context>[];
+export interface Settings<State, Action extends ReduxAction<string>, LoadContext> {
+	routes: RouteWithOrWithoutCodeSplit<State, Action, LoadContext>[];
 	reducers?: Record<string, Reducer>;
 	reduxMiddleware?: Middleware[];
 	reduxStoreEnhancers?: StoreEnhancer[];
@@ -73,7 +73,7 @@ export interface Settings<State, Action extends ReduxAction<string>, Context> {
     onError?: (error: Error, parameters: {
     	location: Location,
     	url: string,
-    	redirect: (to: string | Location) => void,
+    	redirect: (to: LocationInput) => void,
     	dispatch: Dispatch<Action>,
     	useSelector: TypedUseSelectorHook<State>
     }) => boolean | undefined,
@@ -90,7 +90,7 @@ export interface Settings<State, Action extends ReduxAction<string>, Context> {
 	onLoadError?(error: Error, parameters: {
 		location: Location,
 		url: string,
-		redirect: (to: string | Location) => void,
+		redirect: (to: LocationInput) => void,
 		dispatch: Dispatch<Action>,
 		useSelector: TypedUseSelectorHook<State>,
 		server: boolean
@@ -107,9 +107,91 @@ export interface Settings<State, Action extends ReduxAction<string>, Context> {
 // export interface Location extends LocationDescriptor {
 // 	origin: string;
 // }
-export interface Location {
-	origin: string;
+
+// By defaul, `farce` and `found` use a location object
+// that doesn't have a `query` object and has a `search` string instead.
+// This library adds a `queryMiddleware` under the hood which parses `string` into `query`.
+export interface LocationBasic {
+	// `useNavigationLocation()` returns a `location` without the `origin` field.
+	// origin: string;
 	pathname: string;
 	search?: string;
 	hash?: string;
 }
+
+export interface Location {
+	// * `POP` for the initial location, or for any "back"/"forward"/"goto in history" navigation.
+	// * `PUSH` when navigating to a new location.
+	// * `REPLACE` when "redirecting" to a location via DOM API `history.replaceState()`.
+	action: 'POP' | 'PUSH' | 'REPLACE';
+
+	// `origin` is added by this library.
+	// Example: "http://localhost:1234".
+	origin: string;
+
+	// Mimicks the `pathname` part of the DOM `location`:
+	// everything after the `/` sign, including the `/` sign itself.
+	// Is always non-empty.
+	pathname: string;
+
+	// URL query parameters object.
+	// `query` is parsed by `queryMiddleware` which is added by this library
+	// into the list of middlewares when creating a Redux `store`.
+	query: Record<string, string>;
+
+	// Mimicks the `search` part of the DOM `location`:
+	// everything after the `?` sign, including the `?` sign itself.
+	// `search` seems to always be present.
+	// When there's no "hash" part of the URL, it's just an empty string `""`.
+	search: string;
+
+	// Mimicks the `hash` part of the DOM `location`:
+	// everything after the `#` sign, including the `#` sign itself.
+	// `hash` seems to always be present.
+	// When there's no "hash" part of the URL, it's just an empty string `""`.
+	hash: string;
+
+	// Mimicks the `href` property of the DOM `location`:
+	// The full location URL
+	href: string;
+
+	// `delta` is `+1` / `-1` / etc in case of a "goto in history" navigation like "back"/"forward".
+	// In other circumstances, it's gonna be `0`.
+	delta: number;
+
+	// Example: "1234".
+	port: string;
+
+	// Example: "http:".
+	protocol: string;
+
+	// I'd guess, this mimicks the `state` argument that was passed to DOM functions
+	// `history.pushState()` or `history.replaceState()`.
+	// https://developer.mozilla.org/en-US/docs/Web/API/History_API
+	// For example, they seem to store the scroll position in `location.state`.
+	// My guess would be that when navigating not to a `location: string` but to a
+	// `location: object`, one could pass the `state` as part of the `location`.
+	state?: any;
+
+	// The initial location has `key: undefined`.
+	// Any subsequent locations have a (supposedly unique? not guaranteed?) key.
+	// Example: "xdgl89:0".
+	key?: string;
+
+	// The index of the current location in the DOM (navigation) `history`.
+	// Is `0` for the initial location.
+	// Gets incremented for any subsequent "PUSH" navigation.
+	// Doesn't get incremented for "REPLACE" navigation.
+	index: number;
+}
+
+// The `location` argument that navigation functions accept.
+export type LocationInput = string | {
+	pathname: string;
+	// One could pass URL query parameters either in the form of a `search` string
+	// or in the form of a `query` object.
+	query?: Record<string, string>;
+	search?: string;
+	hash?: string;
+	state?: any;
+};